refactor(goals): split GoalDiveDeep into orchestrator + chart + table (step 12)
Extract LifecycleChart and LifecycleTable into separate components.
GoalDiveDeep now orchestrates state and delegates rendering.
